refactor: consolidate logger initialization, move logger files to internal/helpers, and update references
