+++ cmd/root.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/rozdolsky33/ocloud/cmd/configs"
-	"github.com/rozdolsky33/ocloud/cmd/helpers"
+	"github.com/rozdolsky33/ocloud/internal/helpers"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"os"
